[fix] vm

+++ itcloud/src/main/java/com/itinfo/itcloud/model/create/VmSystemVo.java
@@ -3,13 +3,14 @@ package com.itinfo.itcloud.model.create;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.ToString;
+import org.ovirt.engine.sdk4.types.InstanceType;
 
 @Getter @Builder @ToString
 public class VmSystemVo {
     // 시스템
-    private int memorySize;         // 메모리 크기
-    private int memoryMax;          // 최대 메모리
-    private int memoryActual;       // 할당할 실제 메모리
+    private long memorySize;         // 메모리 크기
+    private long memoryMax;          // 최대 메모리
+    private long memoryActual;       // 할당할 실제 메모리
 
     // 고급 매개변수
     private int vCpuCnt;            // 총 가상 CPU
@@ -20,6 +21,8 @@ public class VmSystemVo {
     private String userCpu;         // 사용자 정의 CPU
     private String userVersion;     // 사용자 정의 호환 버전
 
-    private String instanceType;    // 인스턴스 유형
+    // 인스턴스 유형이 바뀌면 가상 소켓과 메모리크기가 변함
+    // https://192.168.0.80/ovirt-engine/api/instancetypes
+    private String instanceType;    // 인스턴스 유형 (none, large, medium, small, tiny, xlarge)
     private String timeOffset;      // 하드웨어 클럭의 시간 오프셋 기본값으로 하면됨 greenwich standard time, kst
 }
